icon fix, refactoring
