+++ components/bl-snackbar-component/src/components/icons/index.js
@@ -2,3 +2,4 @@ export * from './error'
 export * from './info'
 export * from './success'
 export * from './warning'
+export * from './close'
