+++ components/bl-snackbar-component/src/helpers/type.js
